webapp/requirements.txt
# Webapp dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
itsdangerous==2.2.0
jinja2==3.1.4
